::Input FFI
