+++ 3DRadSpace/Engine3DRadSpace/Input/Keyboard.cpp
@@ -1,5 +1,4 @@
 #include "Keyboard.hpp"
-#include "Keyboard.h"
 
 using namespace Engine3DRadSpace::Input;
 
@@ -74,28 +73,4 @@ KeyState Keyboard::operator[](const Key k)
             return KeyState::Up;
     }
     return KeyState::Down;
-}
-
-bool E3DRSP_Keyboard_IsKeyDown(E3DRSP_Keyboard* kb, E3DRSP_Key key)
-{
-    if (kb == nullptr) return false;
-    return (reinterpret_cast<Keyboard*>(kb))->IsKeyDown(static_cast<Key>(key));
-}
-
-bool E3DRSP_Keyboard_IsKeyUp(E3DRSP_Keyboard* kb, E3DRSP_Key key)
-{
-    if (kb == nullptr) return false;
-    return (reinterpret_cast<Keyboard*>(kb))->IsKeyUp(static_cast<Key>(key));
-}
-
-bool E3DRSP_Keyboard_IsAnyKeyDown(E3DRSP_Keyboard* kb)
-{
-    if (kb == nullptr) return false;
-    return (reinterpret_cast<Keyboard*>(kb)->IsAnyKeyDown());
-}
-
-E3DRSP_KeyState E3DRSP_Keyboard_At(E3DRSP_Keyboard* kb, E3DRSP_Key key)
-{
-    if (kb == nullptr) return E3DRSP_KeyState_Down;
-    return static_cast<E3DRSP_KeyState>((reinterpret_cast<Keyboard*>(kb))->operator[](static_cast<Key>(key)));
 }
